Move flow parser definition to own module to resolve circular import error
