+++ OTAnalytics/application/use_cases/load_otflow.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from typing import Iterable
 
-from OTAnalytics.application.datastore import FlowParser
+from OTAnalytics.application.parser.flow_parser import FlowParser
 from OTAnalytics.application.use_cases.event_repository import ClearAllEvents
 from OTAnalytics.application.use_cases.flow_repository import (
     AddFlow,
