+++ tests/OTAnalytics/application/use_cases/test_load_otflow.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock, call
 
 import pytest
 
-from OTAnalytics.application.datastore import FlowParser
+from OTAnalytics.application.parser.flow_parser import FlowParser
 from OTAnalytics.application.use_cases.event_repository import ClearAllEvents
 from OTAnalytics.application.use_cases.flow_repository import (
     AddFlow,
