+++ tests/OTAnalytics/plugin_ui/test_cli.py
@@ -24,11 +24,12 @@ from OTAnalytics.application.config import (
     DEFAULT_NUM_PROCESSES,
     DEFAULT_TRACK_FILE_TYPE,
 )
-from OTAnalytics.application.datastore import FlowParser, TrackParser, VideoParser
+from OTAnalytics.application.datastore import TrackParser, VideoParser
 from OTAnalytics.application.eventlist import SceneActionDetector
 from OTAnalytics.application.logger import DEFAULT_LOG_FILE
 from OTAnalytics.application.parser.cli_parser import CliArguments, CliParseError
 from OTAnalytics.application.parser.config_parser import ConfigParser
+from OTAnalytics.application.parser.flow_parser import FlowParser
 from OTAnalytics.application.run_configuration import RunConfiguration
 from OTAnalytics.application.state import TracksMetadata, VideosMetadata
 from OTAnalytics.application.use_cases.apply_cli_cuts import ApplyCliCuts
